requirements Clayde.txt
python-telegram-bot==20.7
openai==1.12.0
pyswisseph==2.10.3.2
matplotlib==3.8.2
pytz==2024.1
timezonefinder==6.2.0
geopy==2.4.1
APScheduler==3.10.4
tenacity==8.2.3
python-dotenv==1.0.0